version1_backup/requirements.txt
appdirs==1.4.4
asttokens==3.0.0
beautifulsoup4==4.13.4
blinker==1.9.0
branca==0.8.1
bs4==0.0.2
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
cloudscraper==1.2.71
colorama==0.4.6
contourpy==1.3.2
cssselect==1.3.0
cycler==0.12.1
decorator==5.2.1
et_xmlfile==2.0.0
executing==2.2.0
fake-useragent==2.2.0
Flask==3.1.1
folium==0.20.0
fonttools==4.58.4
geographiclib==2.0
geopandas==1.1.0
geopy==2.4.1
gunicorn==23.0.0
haversine==2.9.0
idna==3.10
importlib_metadata==8.7.0
ipython==9.3.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
lxml==5.4.0
lxml_html_clean==0.4.2
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
networkx==3.5
numpy==2.3.0
openpyxl==3.1.5
OWSLib==0.34.1
packaging==25.0
pandas==2.3.0
parse==1.20.2
parso==0.8.4
pillow==11.2.1
prompt_toolkit==3.0.51
pure_eval==0.2.3
pyaml==25.7.0
pyee==11.1.1
Pygments==2.19.1
pyogrio==0.11.0
pyparsing==3.2.3
pyppeteer==2.0.0
pyproj==3.7.1
pyquery==2.0.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-html==0.10.0
requests-toolbelt==1.0.0
scikit-learn==1.7.0
scikit-optimize==0.10.2
scipy==1.16.0
shapely==2.1.1
six==1.17.0
soupsieve==2.7
stack-data==0.6.3
threadpoolctl==3.6.0
tqdm==4.67.1
traitlets==5.14.3
typing_extensions==4.14.1
tzdata==2025.2
urllib3==1.26.20
w3lib==2.3.1
wcwidth==0.2.13
websockets==10.4
Werkzeug==3.1.3
xyzservices==2025.4.0
zipp==3.23.0
